Handle Literal+CharClass, this makes `{x!~"host1|host2"}` work correctly (#72)
I suspect this is the limit of the approach this currently takes, if
we want to support the full possibilities of fixed string extraction
this needs to take a different approach, like actually running the
regexp, not just looking for patterns (i.e. use Prog, not the Regexp
syntax tree).