Add distributed tracing (OpenTelemetry, Jaeger exporter)
This commit adds basic distributed tracing by instrumenting the grpc server and
the http client with the OpenTelemetry framework. This is intended as an initial
stepping stone to add "deeper" specific Geras traces in the future.

With this commit, a request can be traced from Prometheus through
thanos-remote-read (or Thanos) through Geras and exported to Jaeger.