Support otlp trace exporter and current otel libraries (#102)
Configuration diverges from the environment specification, https://opentelemetry.io/docs/reference/specification/sdk-environment-variables/#exporter-selection, in that the default OTEL_TRACES_EXPORTER is none. This is to provide compatibility with the previous default and to avoid surprising infrastructure dependencies.

Other changes in configuration are due to changes in the OpenTelemetry Jaeger exporter as of v0.20.0: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v0.20.0.

As it is deprecated, we will remove support for the Jaeger exporter in a future version. This is intended as a migration-enabling version to allow people to convert from jaeger to the otlp exporter.